Optimize State.observe_removes
observe_removes was one of the slowest parts of the State merge
function. Each time observe_removes was called, we did a full ets table
scan on the values ets table and copied all the elements except from
local replica and then filtered them one by one. This was especially
slow when there were multiple nodes each with many tracked processes.
E.g. given 4 replicas each with 25K records, we had to scan all 100K
records and then copied 75K from ets for additional processing.

We had to go through that each time when there was a non-empty delta,
even if we didn't need to remove any records.

This allows us to avoid full table scan and copying a lot of records
when we're merging a delta. I did not change the behaviour when a
regular state (non delta) is merged.