Allow triggering error in shard
This is for debugging only and won't be used in production.

When processing state update, read value from `:trigger_shard_error_for_topic`
persistent term, and trigger an error if topic prefix matches with it.
Also reset the flag each time right before triggering an error. This way
we can observe how the system fails and/or recovers when a shard has blown up once.

The trigger can be set in application (e.g. via HTTP request).