Make ets insert and delete more strict
`:ets.insert` returns true when a record was updated or insterted. We
know that we only want to insert, so `:ets.insert_new` makes more sense,
which returns true only when a record was inserted.

Instead of using `:ets.delete` and `:ets.match_delete` we can use
`select_delete` which returns the number of rows affected. If it
something else that 1 then we know we've encountered a problem.